+++ packages/fields/CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## Unreleased
 
+## 0.1.0 (2024-10-03)
+
 Initial release.